altera tipo da chave primaria da UserModel

+++ src/main/java/validationpojos/springvalidation/repository/UserRepository.java
@@ -8,6 +8,6 @@ import validationpojos.springvalidation.model.UserModel;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserModel, Integer> {
+public interface UserRepository extends JpaRepository<UserModel, Long> {
     Optional<UserModel> findByLogin(String login);
 }
